# doc/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "华农掌中行 API",
    "version": "1.0.0",
    "description": "由 back/server.js 与 back/routes/*.js 导出的接口定义（2026-08-14）。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地后端（以 PORT 环境变量为准）"
    }
  ],
...
  "paths": {
...
    "/api/favorites/add": {
      "post": {
        "tags": [
          "收藏"
        ],
        "summary": "添加收藏",
        "operationId": "post_api_favorites_add",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "number"
                  },
                  "poiId": {
                    "type": "number"
                  }
                },
                "required": [
                  "userId",
                  "poiId"
                ]
              }
            }
          }
        }
      }
    },
    "/api/favorites/remove": {
      "post": {
        "tags": [
          "收藏"
        ],
        "summary": "取消收藏",
        "operationId": "post_api_favorites_remove",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "number"
                  },
                  "poiId": {
                    "type": "number"
                  }
                },
                "required": [
                  "userId",
                  "poiId"
                ]
              }
            }
          }
        }
      }
    },
    "/api/favorites/list": {
      "get": {
        "tags": [
          "收藏"
        ],
        "summary": "收藏列表",
        "operationId": "get_api_favorites_list",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
    "/api/favorites/check": {
      "get": {
        "tags": [
          "收藏"
        ],
        "summary": "检查收藏状态",
        "operationId": "get_api_favorites_check",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "poiId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
...
    "/api/admin/flower/routes/list": {
      "get": {
        "tags": [
          "后台业务管理"
        ],
        "summary": "管理端赏花路线列表",
        "operationId": "get_api_admin_flower_routes_list",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 10
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/admin/flower/routes/add": {
      "post": {
        "tags": [
          "后台业务管理"
        ],
        "summary": "新增赏花路线",
        "operationId": "post_api_admin_flower_routes_add",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "number"
                  },
                  "distance": {
                    "type": "number"
                  },
                  "difficulty": {
                    "type": "string"
                  },
                  "spots": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "description": {
                    "type": "string"
                  },
                  "best_time": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "highlights": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "tips": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "userToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "用户 JWT，直接传 token"
      },
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "管理员 JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": [
          "code"
        ]
      }
    }
  }
}